Safari-609.1.11.1/LayoutTests/css3/flexbox/flex-column-relayout-assert.html

<!DOCTYPE html>
<link href="resources/flexbox.css" rel="stylesheet">
<style>
.flexbox {
    background-color: green;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.flexbox')">
<div id=log></div>

<p>You should see a green rectangle, 20px high.</p>

<div class="flexbox column" data-expected-height="40">
  <div id="child" data-expected-height="40"></div>
</div>

<script>
document.getElementById('child').offsetHeight;
document.getElementById('child').style.padding = "20px";
</script>
